refactor(cli): resolve review findings — no cross-package source import, shared story-paths helper
- Extract detectLanguage/detectIncompatiblePackageVersions into core
  (storybook/internal/cli) and delegate from ProjectTypeService, so core
  no longer imports source from create-storybook (which depends on core).
- Move getStoriesPathsFromConfig into core-server as the single shared
  copy; cli-storybook re-exports it and the ai module imports it, instead
  of two verbatim duplicates.
