+++ code/core/src/cli/ai/index.ts
@@ -7,8 +7,7 @@ import { logger } from 'storybook/internal/node-logger';
 import { telemetry } from 'storybook/internal/telemetry';
 import { SupportedLanguage } from 'storybook/internal/types';
 
-import { ProjectTypeService } from '../../../../lib/create-storybook/src/services/ProjectTypeService.ts';
-
+import { detectLanguage } from '../detectLanguage.ts';
 import { getStorybookData } from './storybook-data.ts';
 import { getAiSetupMarkdownOutput } from './setup-prompts/index.ts';
 import type { ProjectInfo, AiSetupOptions } from './types.ts';
@@ -35,8 +34,7 @@ export async function aiSetup(options: AiSetupOptions): Promise<void> {
       ? parseMajorVersion(data.versionInstalled)
       : undefined;
 
-    const projectTypeService = new ProjectTypeService(data.packageManager);
-    const detectedLanguage = await projectTypeService.detectLanguage();
+    const detectedLanguage = await detectLanguage(data.packageManager);
     const language = detectedLanguage === SupportedLanguage.TYPESCRIPT ? 'ts' : 'js';
 
     const needsUserOnboarding = await cache.get<boolean>('onboarding-pending', false);
