+++ code/lib/cli-storybook/src/util.ts
@@ -1,7 +1,12 @@
 import type { PackageJsonWithDepsAndDevDeps } from 'storybook/internal/common';
-import { HandledError, JsPackageManager, normalizeStories } from 'storybook/internal/common';
+import { HandledError, JsPackageManager } from 'storybook/internal/common';
 import { getProjectRoot, isSatelliteAddon, versions } from 'storybook/internal/common';
-import { StoryIndexGenerator, experimental_loadStorybook } from 'storybook/internal/core-server';
+import {
+  experimental_loadStorybook,
+  getStoriesPathsFromConfig,
+} from 'storybook/internal/core-server';
+
+export { getStoriesPathsFromConfig };
 import { logTracker, logger, prompt } from 'storybook/internal/node-logger';
 import {
   UpgradeStorybookToLowerVersionError,
@@ -775,48 +780,4 @@ export const getEvaluatedStoryPaths = async (
     configDir,
     workingDir,
   });
-};
-
-/**
- * Gets story file paths from a Storybook configuration directory
- *
- * @example
- *
- * ```typescript
- * const storiesPaths = await getStoriesPathsFromConfigWithoutEvaluating({
- *   stories: ['src\/**\/*.stories.tsx'],
- *   configDir: '/path/to/.storybook',
- *   workingDir: '/path/to/project',
- * });
- * ```
- */
-export const getStoriesPathsFromConfig = async ({
-  stories,
-  configDir,
-  workingDir,
-}: {
-  stories: StorybookConfigRaw['stories'];
-  configDir: string;
-  workingDir: string;
-}) => {
-  if (stories.length === 0) {
-    return [];
-  }
-
-  const normalizedStories = normalizeStories(stories, {
-    configDir,
-    workingDir,
-  });
-
-  const matchingStoryFiles = await StoryIndexGenerator.findMatchingFilesForSpecifiers(
-    normalizedStories,
-    workingDir,
-    true
-  );
-
-  const storiesPaths = matchingStoryFiles.flatMap(([specifier, cache]) => {
-    return StoryIndexGenerator.storyFileNames(new Map([[specifier, cache]]));
-  });
-
-  return storiesPaths;
 };
